requirements.txt
google-generativeai==0.8.3
serpapi==0.1.5
transformers==4.36.0
torch==2.1.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
nltk==3.8.1